twitter_analysis/requirements.txt
flask==2.0.1
tweepy==4.12.1
python-dotenv==0.19.0
pymongo==4.3.3
werkzeug==2.3.7 